utils: Move Serial from wayland
