+++ smallvil/src/handlers/xdg_shell.rs
@@ -14,14 +14,13 @@ use smithay::{
             Resource,
         },
     },
-    utils::Rectangle,
+    utils::{Rectangle, Serial},
     wayland::{
         compositor::with_states,
         shell::xdg::{
             PopupSurface, PositionerState, ToplevelSurface, XdgShellHandler, XdgShellState,
             XdgToplevelSurfaceRoleAttributes,
         },
-        Serial,
     },
 };
