+++ src/wayland/seat/keyboard.rs
@@ -15,7 +15,7 @@ use crate::{
         Seat, SeatHandler, SeatState,
     },
     utils::IsAlive,
-    wayland::Serial,
+    utils::Serial,
 };
 
 impl<D: SeatHandler + 'static> KeyboardHandle<D> {
