+++ src/wayland/xdg_activation/mod.rs
@@ -72,7 +72,7 @@ use wayland_server::{
 
 use rand::distributions::{Alphanumeric, DistString};
 
-use crate::wayland::Serial;
+use crate::utils::Serial;
 
 mod dispatch;
